move ProxyParameters to public API
