+++ core/src/main/java/io/grpc/ProxyDetector.java
@@ -16,7 +16,6 @@
 
 package io.grpc;
 
-import io.grpc.internal.ProxyParameters;
 import java.io.IOException;
 import java.net.SocketAddress;
 import javax.annotation.Nullable;
