+++ core/src/main/java/io/grpc/internal/ProxySocketAddress.java
@@ -18,6 +18,7 @@ package io.grpc.internal;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
+import io.grpc.ProxyParameters;
 import java.net.SocketAddress;
 
 /**
